refactor(agent): share routing policy contract
